BUG: Updated these internal tests to reflect API changes

+++ Utilities/MetaIO/tests/testMeta2Object.cxx
@@ -16,7 +16,7 @@ int main(int argc, char **argv)
   tObj.Position(0, 1);
   tObj.Position(1, 2);
   tObj.Position(2, 3);
-  float orient[9];
+  double orient[9];
   int i;
   for(i=0; i<9; i++)
     {
